requirements.txt
# Core
fastapi==0.128.0
uvicorn[standard]==0.40.0
gunicorn==23.0.0
pydantic==2.12.5

# ML / Vision
torch==2.9.1
torchvision==0.24.1
transformers==4.57.6
Pillow==12.1.0

# HTTP / Utils
httpx==0.28.1
numpy==2.4.1
python-multipart==0.0.21

# Development
pytest==9.0.2
pytest-asyncio==1.3.0
